feat: Types for Axios
