+++ src/types/GenericResponse.tsx
@@ -1,0 +1,4 @@
+interface GenericResponse {
+  status: string;
+  message: string;
+}
