+++ src/types/LoginInput.tsx
@@ -1,0 +1,4 @@
+interface LoginInput {
+  email: string;
+  password: string;
+}
