+++ src/types/User.tsx
@@ -1,0 +1,8 @@
+export interface IUser {
+  email: string;
+  firstname: string;
+  lastname: string;
+  phonenumber: number;
+  password: string;
+  confirmPassword: string;
+}
